llama2-7b-chat-hf-GGML/requirements.txt
ctransformers
transformers
gradio_client==0.2.7